refactor(executor): extract CQL execution engine into injectable executor module
Move non-interactive CQL execution out of main.rs into a new src/executor.rs
module. All execute_* functions now accept a `&mut dyn Write` writer instead
of writing directly to stdout, enabling in-process execution with captured
output. main.rs delegates to executor::* passing `&mut io::stdout()`.

Add `run_cql_in_process(host, port, keyspace, cql) -> Result<String>` to
lib.rs so integration tests can call the full execution pipeline without
spawning a subprocess, making the code paths visible to cargo tarpaulin.
